+++ CHANGELOG.md
@@ -21,9 +21,10 @@ v2.6.0 [unreleased]
 
 ### Features
 - [#530](https://github.com/latera/homs/pull/530) Customize fonts in HBW.
-- [#534](https://github.com/latera/homs/pull/534), [#536](https://github.com/latera/homs/pull/536) Add custom top/bottom descriptions and placeholder for components:
+- [#534](https://github.com/latera/homs/pull/534), [#536](https://github.com/latera/homs/pull/536) [#543](https://github.com/latera/homs/pull/543) Add custom top/bottom descriptions and placeholder for components:
   * string
   * select
+  * checkbox
 
   Now you can specify some additional description placed over/under the input for fields listed above with option:
   ```